Feature/topup page fe

+++ ewallet-server/src/main/java/com/berijalan/ewallet/repository/MerchantRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface MerchantRepository extends JpaRepository<Merchant, Integer> {
+public interface MerchantRepository extends JpaRepository<Merchant, Long> {
     Optional<Merchant> findByName(String name);
 }
